Splitting up serializer into types/serializer/visitors. Adding Flow annotations as needed at module boundaries.
