+++ src/scripts/generate-sourcemaps-test.js
@@ -9,7 +9,7 @@
 
 /* @flow */
 
-import Serializer from "../serializer.js";
+import Serializer from "../serializer/index.js";
 import invariant from "../invariant.js";
 
 let chalk = require("chalk");
